Complete.tsxとRecommended.tsxでデータベースからデータを取得するように変更した
